Add files via upload
Problem A - Many Equal Substrings from Codeforces Round #506.
I solved this problem by starting at index 1 of a given string. I then took a substring from the current index to the end of the string, checking to see if the given string started also with the substring. If so, the program would then record the current index as an integer and break the loop. It then would take the original string and append the substring found above k-1 times.